More tests running on Windows

+++ airflow/operators/bash_operator.py
@@ -88,6 +88,9 @@ class BashOperator(BaseOperator):
         Execute the bash command in a temporary directory
         which will be cleaned afterwards
         """
+        if os.name == 'nt' and self.execution_timeout:
+            raise NotImplementedError("Cannot use execution timeouts with BashOperator on Windows")
+
         self.log.info('Tmp dir root location: \n %s', gettempdir())
 
         # Prepare env for child process.
@@ -105,45 +108,47 @@ class BashOperator(BaseOperator):
         self.lineage_data = self.bash_command
 
         with TemporaryDirectory(prefix='airflowtmp') as tmp_dir:
-            with NamedTemporaryFile(dir=tmp_dir, prefix=self.task_id) as tmp_file:
+            tmp_file_name = os.path.join(tmp_dir, "bash_operator.cmd")
+            with open(tmp_file_name, "w") as tmp_file:
                 tmp_file.write(bytes(self.bash_command, 'utf_8'))
-                tmp_file.flush()
-                script_location = os.path.abspath(tmp_file.name)
-                self.log.info('Temporary script location: %s', script_location)
+            script_location = os.path.abspath(tmp_file.name)
+            self.log.info('Temporary script location: %s', script_location)
 
-                def pre_exec():
-                    # Restore default signal disposition and invoke setsid
-                    for sig in ('SIGPIPE', 'SIGXFZ', 'SIGXFSZ'):
-                        if hasattr(signal, sig):
-                            signal.signal(getattr(signal, sig), signal.SIG_DFL)
-                    os.setsid()
+            def pre_exec():
+                # Restore default signal disposition and invoke setsid
+                for sig in ('SIGPIPE', 'SIGXFZ', 'SIGXFSZ'):
+                    if hasattr(signal, sig):
+                        signal.signal(getattr(signal, sig), signal.SIG_DFL)
+                os.setsid()
 
-                self.log.info('Running command: %s', self.bash_command)
-                sub_process = Popen(
-                    ['bash', tmp_file.name],
-                    stdout=PIPE,
-                    stderr=STDOUT,
-                    cwd=tmp_dir,
-                    env=env,
-                    preexec_fn=pre_exec if os.name != 'nt' else None)
+            self.log.info('Running command: %s', self.bash_command)
+            sub_process = Popen(
+                ['bash', tmp_file_name] if os.name != 'nt' else [tmp_file_name],
+                stdout=PIPE,
+                stderr=STDOUT,
+                cwd=tmp_dir,
+                env=env,
+                preexec_fn=pre_exec if os.name != 'nt' else None)
 
-                self.sub_process = sub_process
+            self.sub_process = sub_process
 
-                self.log.info('Output:')
-                line = ''
-                for raw_line in iter(sub_process.stdout.readline, b''):
-                    line = raw_line.decode(self.output_encoding).rstrip()
-                    self.log.info(line)
+            self.log.info('Output:')
+            line = ''
+            for raw_line in iter(sub_process.stdout.readline, b''):
+                line = raw_line.decode(self.output_encoding).rstrip()
+                self.log.info(line)
 
-                sub_process.wait()
+            sub_process.wait()
 
-                self.log.info('Command exited with return code %s', sub_process.returncode)
+            self.log.info('Command exited with return code %s', sub_process.returncode)
 
-                if sub_process.returncode:
-                    raise AirflowException('Bash command failed')
+            if sub_process.returncode:
+                raise AirflowException('Bash command failed')
 
         return line
 
     def on_kill(self):
+        if os.name == 'nt':
+            return
         self.log.info('Sending SIGTERM signal to bash process group')
         os.killpg(os.getpgid(self.sub_process.pid), signal.SIGTERM)
